refactor(extension)!: Transform the extension as a "bundle"
