+++ package.json
@@ -2,31 +2,46 @@
   "name": "directus-extension-generate-types",
   "version": "0.4.2",
   "license": "MPL-2.0",
-  "keywords": [
-    "directus",
-    "directus-extension",
-    "directus-custom-module"
-  ],
-  "directus:extension": {
-    "type": "module",
-    "path": "dist/index.js",
-    "source": "src/index.ts",
-    "host": "^9.0.0-rc.95",
-    "hidden": false
-  },
-  "scripts": {
-    "build": "directus-extension build",
-    "watch": "directus-extension build --watch",
-    "dev": "./dev.sh"
-  },
-  "devDependencies": {
-    "@directus/extensions-sdk": "9.18.1",
+	"type": "module",
+	"description": "Generate types for use in TypeScript, Python or for OpenAPI usage",
+	"icon": "extension",
+	"keywords": [
+		"directus",
+		"directus-extension",
+		"directus-custom-bundle"
+	],
+	"directus:extension": {
+		"type": "bundle",
+		"path": {
+			"app": "dist/app.js",
+			"api": "dist/api.js"
+		},
+		"host": "^10.1.5",
+    "hidden": false,
+		"entries": [
+			{
+				"type": "module",
+				"name": "generate-types-admin-module",
+				"source": "src/generate-types-admin-module/index.ts"
+			}
+		]
+	},
+	"scripts": {
+		"build": "directus-extension build",
+		"dev": "directus-extension build -w --no-minify",
+		"dev:project": "sh ./dev.sh",
+		"link": "directus-extension link",
+		"add": "directus-extension add"
+	},
+	"devDependencies": {
+		"@directus/extensions-sdk": "10.1.5",
+		"@types/node": "^20.4.0",
     "@types/prismjs": "^1.26.0",
     "axios": "^0.27.2",
     "sass": "^1.55.0",
-    "typescript": "^4.8.4",
-    "vue": "^3.2.40"
-  },
+		"typescript": "^5.1.6",
+		"vue": "^3.3.4"
+	},
   "dependencies": {
     "@directus/shared": "^9.18.1",
     "prismjs": "^1.29.0"
