+++ src/generate-types-admin-module/index.ts
