+++ src/generate-types-admin-module/lib/console.ts
