+++ src/generate-types-admin-module/routes/py.vue
@@ -13,8 +13,8 @@
 <script lang="ts">
 import NavbarComponent from "../components/navigation.vue";
 import CodeComponent from "../components/code.vue";
-import generatePyTypes from "lib/generateTypes/py";
-import languages from "lib/languages";
+import generatePyTypes from "../lib/generateTypes/py";
+import languages from "../lib/languages";
 
 export default {
   components: { NavbarComponent, CodeComponent },
