+++ src/generate-types-admin-module/utils/use-clipboard.ts
