+++ src/generate-types-admin-module/vue-shims.d.ts
